Move functions into ImageryCoverage class
